added websocket server...
this required to refactor some of the code to avoid circular dependencies. as a result the CLI code could be greatly reduced and is now pretty simple
* pytaglib was added as requirement, which is now used to keep meta data from the source file and to add meta data with the RVC generation flags to the output file (this could/should probably also be added to batch conversion)
* cli interface was made executable so that it can be run from CLI on linux/mac systems
