+++ requirements.txt
@@ -43,3 +43,4 @@ onnxruntime-gpu; sys_platform != 'darwin'
 torchcrepe==0.0.20
 fastapi==0.88
 ffmpy==0.3.1
+pytaglib>=2.0.0
